Add login and role: - created servlet Login and page; - added role of user; - added rollbakc() for method delete; - changed PageGenerator

+++ src/main/java/com/rk/web/servlet/DeleteServlet.java
@@ -1,20 +1,41 @@
 package com.rk.web.servlet;
 
 import com.rk.ServiceLocator;
+import com.rk.domain.User;
 import com.rk.service.GlassesService;
 
+import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.util.Map;
 
 public class DeleteServlet extends HttpServlet {
-    private final GlassesService glassesService = ServiceLocator.getBean(GlassesService.class);
+    private GlassesService glassesService;
+    private Map<String, User> cookieTokens;
+
+    public DeleteServlet() {
+        this.glassesService = ServiceLocator.getBean(GlassesService.class);
+        this.cookieTokens = ServiceLocator.getBean(Map.class);
+    }
 
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        long id = Long.parseLong(request.getParameter("id"));
-        glassesService.deleteById(id);
-        response.sendRedirect("");
+        Cookie[] cookies = request.getCookies();
+        if (cookies != null) {
+            for (Cookie cookie : cookies) {
+                if (cookie.getName().equals("user-token")) {
+                    User user = cookieTokens.get(cookie.getValue());
+                    if (user != null && user.getRole().getUserRole().equals("ADMIN")) {
+                        long id = Long.parseLong(request.getParameter("id"));
+                        glassesService.deleteById(id);
+                        response.sendRedirect("");
+                        return;
+                    }
+                }
+            }
+            response.sendRedirect("/login");
+        }
     }
 }
